biometry_kernel/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Vaez Vladimir <Skader-kot@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.7.1"
fastapi = "^0.110.2"
pillow = "^10.3.0"
asyncpg = "^0.29.0"
sqlalchemy = "^2.0.29"
opencv-python = "^4.9.0.80"
python-multipart = "^0.0.9"
torch = "2.3.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
